VIT-6747: Menstrual Cycle - sync pt.2 (#157)

+++ Sources/VitalCore/Core/Client/Data Models/Gets/MenstrualCycle.swift
@@ -225,11 +225,11 @@ extension MenstrualCycle {
 
   public struct BasalBodyTemperatureEntry: Equatable, Codable {
     public let date: GregorianCalendar.Date
-    public let value: Decimal
+    public let value: Double
 
     public init(
       date: GregorianCalendar.Date,
-      value: Decimal
+      value: Double
     ) {
       self.date = date
       self.value = value
